libcupsfilters: Don't call cupsRasterParseIPPOptions() if we have PPD
In a filter function, when we have a PPD file we should not call
cupsRasterParseIPPOptions() as this function guesses the meaning of
job options/attributes from their names.

This can easily cause problems, for example if we supply
"print-color-mode=monochrome" and the function chooses the wromg
monochrome color space, one not supported according to the PPD and so
not supported by the printer or the (rasterto...) driver.

Therefore we remove these calls and trust the PPD's options.

In retro-fitting Printer Applications job IPP attributes are
associated with appropriate PPD options anyway.