libcupsfilters: Corrected color management API
To the color management functions of cupsfilters/colormanager.c got a
pointer to a filter data record (filter_data_t) added to their
parameter lists to convey a log function to them, so that they log
into the caller's log function and not to stderr any more.

It was not taken into account that the data structure also contains
the print queue name and so the print queue name was continued to be
explicitly supplied as the second parameter.

This is not needed any more and so we have changed the parameter lists
of the functions, not explicitly taking the print queue name any more
and the functions use the print queue name from the data structure.

Also fixed some bugs:

- Do not use getenv("PRINTER") in filter functions. The print queue
  name is part of the filter_data_t data structure.

- In foomatic-rip make sure that the correct print queue name gets
  added to the data structure before the first call of a color
  management function happens. Also make sure that color management
  also works in spooler-less mode.