libcupsfilters, rastertopwg: Adjusted default output formats
If the output format is not supplied as parameter, ghostscript()
defaults to CUPS Raster now, as the ghostscript CUPS filter already
does.

rastertopwg does not support CUPS Raster output, so we changed the
CUPS filter rastertopwg to default to PWG Raster instead.

The rastertopwg() filter function errors out when NULL is supplied as
parameter. We let it also error out now if an unsupported output
format is supplied.