libcupsfilters: Renamed internal functions and data types
Renamed the internal functions and data types according to the rules
of the DEVELOPING.md file of the CUPS source code, to extend this
coding style to cups-filters.

1. Renamed source-file-local functions to all-lowercase with "_"
   separators, also declared all these functions static (= valid only
   for the same source file) if not yet done so. These functions do
   not appear in *.h files.

2. Renamed library-local (non-API, but used in several files of the
   library) to camel-case starting with "_cf...".

3. Renamed data types to all-lowercase with "_" separators, ending
   with "..._e" for enumerations, "..._s" for structs, and "..._t" for
   type definitions.

Also moved the content of the files cupsfilters/image-sgi.h and
cupsfilters/image-sgilib.c into cupsfilters/image-sgi.c and removed
unused code for outputting SGI images, and re-arranged the order of
the functions in the cupsfilters/texttopdf.c source file (to not have
the main function in the middle of the file).