Build system, cfFilterUniversal(), cfFilterPDFToRaster(): libpoppler
Fixed building libcupsfilters with and without libpoppler. By default,
libpoppler is used, to not use it, supply the "--disable-poppler"
option to "./configure".

Without libpoppler the cfFilterUniversal requires Ghostscript to turn
PDF input into any of the Raster formats, built with both
"--disable-poppler" and "--disable-ghostscript", the universal filter
function is not able to rasterize PDF.

The cfFilterPDFToRaster() gets non-functional when building
libcupsfilters without libpoppler. Calling it simply produces an error
message in the log.

In addition, unnecessary "./configure" options for the former
PDF-to-PostScript filter function (has moved to libppd as
ppdFilterPDFToPS) which were forgotten during the separation, are
removed now.

Also the "AC_DEFINE" macro calls in configure.ac got corrected,
setting the macros in config.h really to 1 after testing positive.

And a superfluous "HAVE_CUPS_1_7" macro in cfFilterPDFToRaster() got
removed now.