Add language/locale support to cfCatalog...() API functions
The `cfCatalog...()` set of API functions (from the `cupsfilters/catalog.c` file) only served for getting English human-readable strings for options and choices of options, not translations in a requested language. As the human-readable strings are taken from the translation tables of CUPS and also of IPP printers, language support is trivial, only adding a parameter to supply the desired language (in `xx` or `xx_YY` format). This is done with this commit.

It was primarily needed to get human-readable string and translation support in the Common Print Dialog Backends (CPDB), but could in the future also be used for the PPD file generator in libppd.

Thanks, Gaurav Guleria, for this.