cfImageOpenFP(): Removed leftover HAVE_LIBZ conditionals
In an earlier commit we have removed the dependency on libz from the
build system as there is no explicit dependency on it in
libcupsfilters.

The conditionals for libpng availability (HAVE_LIBPNG) were combined
with HAVE_LIBZ, probably due to internal use of libz by libpng. This
is unnecessary and broke PNG image reading support since HAVE_LIBZ
does not get defined any more by the build system.

So we have simply removed the HAVE_LIBZ conditionals now.