Filter functions: No fallback to default media size when no size specified
Especially when using the cfFilterImageTo...() filter functions, when
not specifying the output page size (via job options/attributes
PageSize, media, media-col, ...) there was no fallback to a default
size (from printer IPP attributes, PPD file via libppd, or US Letter
as last mean) and a broken CUPS Raster or PDF file got produced.

Problems were:

1. In cfGetPageDimensions() presence of a Raster header was considered
a page size specified for the job (by the header's page dimensions)
and the size not being valid (header created from scratch by
cfRasterPrepareHeader() does not contain page dimensions) made the
function reporting an error condition of the job's requested page size
not being available.

To remedy this we consider the page dimensions given in a Raster header
as defaults, so the presence of invalid values cannot trigger an error
condition.

2. In cfRasterPrepareHeader() the fallback to US Letter is positioned
too early, before trying to get values from the Raster header in case
of failure of cfGetPageDimensions(), and also invalid values (zero
dimensions) are accepted here.

Here we do the fallback to US Letter now only after failure of all
means. And we reject zero dimensions in the header.

Now default page dimensions of the printer are always used in case of
jobs with no page size requested, and we fall back to US Letter if we
do not have printer capability information in such a case. In
addition, for PDF or PostScript input, the input page sizes are used
if no size is requested for the job.

Fixes https://github.com/OpenPrinting/libcupsfilters/issues/26