Ignore unsupported resolution values
Since some time the print dialog of Google's Chrome browser send the
attribute "Resolution=96dpi" with its print jobs (job itself sent in
PDF format). This is a resolution which most modern printers do not
support and usually the print filters and drivers should ignore
unsupported values for any option or IPP attribute, so such an
attribute should not harm.

Unfortunately, in some cases the validity of the resolution value is
not checked and so the wrong value makes it through to the actual
rasterization, messing up the printouts as a raster with a much too
low, unsupported resolution is sent to the printer.

This commit solves the problem by actually checking the resolution
values. This is done in the raster_base_header() function which
generates a CUPS/PWG Raster header from scratch, solely based on
printer and job IPP attributes, and the option list supplied to the
filter functions. This function is called by the
cfRasterPrepareHeader() API function which creates a CUPS/PWG Raster
header for the filter functions, either as base for the page headers
in CUPS/PWG/Apple Raster output or to get a data structure with all
important page properties. He we add a check now, by passing
resolutions parsed from the options through the
cfIPPAttrResolutionForPrinter() function. This way wrong values get
ignored and the default value be used instead. If there are no printer
attributes available, any resolution value is accepted.

Fixes #29: https://github.com/OpenPrinting/libcupsfilters/issues/29