Added GitHub workflow for Canonical Open Documentation Academy
OpenPrinting will participate in Canonical's Open Documentation
Academy, as an organization in need of documentation.

Essential for that is that documentation-related issue reports on
OpenPrinting GitHub repositories get forwarded to the GitHub
repository of the Open Documentation Academy:

https://github.com/canonical/open-documentation-academy/

This commit is a first test for an automated workflow to do so. We add
the GitHub workflow suggested by the user @minaelee (Minae Lee) in the
feature request for a GitHub action supporting the automated
forwarding:

https://github.com/canonical/open-documentation-academy/issues/11

The suggested workflow code is rudimentary, just copying the issue
report as the label "ODA" gets added to it. There is no syncing of the
copy with the original later on, for example when the original's
description gets edited, comments added to the original or the
original being closed.

Also this workflow is a significant bunch of code, and so a
maintenance nightmare if deploying on all the > 30 repositories on
OpenPrinting.

Therefore we will test it on libcupsfilters for now and include the
other repositories as soon as there is a GitHub action which is
developed and maintained by the Open Documentation Academy.