Allow building without fontconfig
The texttopdf filter is the only thing using fontconfig.  Add a
configure option to build without fontconfig and don't build this filter
if that is specified.

The default configure is to use fontconfig, so there shoule be no
changes to existing functionality.

Also, fix a minor compiler warning when building without exif.