Allow building without fontconfig (#83)
The texttopdf filter (cfFilterTextToPDF() filter function is the only thing
using fontconfig. 

Added a configure option to build without fontconfig and don't
build this filter if that is specified.

The default configuration is to use fontconfig, so there should be no
changes to existing functionality.

Instead of only defining the cfFilterTextToPDF() filter function if
HAVE_FONTCONFIG is defined, always define this function and just exit
early if there is no fontconfig.

Also, fixed a minor compiler warning when building without exif.